+++ src/constants/alertCode.ts
@@ -1,0 +1,10 @@
+/** 
+ *- CA: Comment Alert
+  - EA: Event Alert
+ */
+const enum AlertCode {
+  CA = 'CA',
+  EA = 'EA',
+}
+
+export default AlertCode;
