+++ src/models/alert/AlertType.ts
@@ -1,3 +1,3 @@
-import { AlertCode } from 'constants/code';
+import AlertCode from 'constants/alertCode';
 
 export type AlertType = AlertCode.CA | AlertCode.EA;
